Math libs rename
Give to some files more specific names
